Remove img fullscreen from Compare
Full screen functionality was added to the System.

It was meant to be on the Simulator, as it exists an image there.
But it was created on the utils package which is shared by both the Simulator and Comparator.

So it was moved to the Simulator js's only.

Next it'll be brought back to the utils package with the exception.